Deprecate the autoloaders, and just use rails initializers for this in the standard way. Gemfile inclusion is sufficient for control. Add reset! hooks to prevent test environment leakage. Autoload the railtie under rails environments. Cleanup the railtie, merging into the main file
