--- lib/sprites/rails/autoload.rb
@@ -1,7 +1,0 @@
-require 'sprites/rails'
-require 'sprites/autoloader'
-
-class Sprites
-  class Railtie
-  end
-end
