--- spec/fixtures/rails_project1/Rakefile
@@ -1,5 +1,0 @@
-require File.expand_path('../config/application', __FILE__)
-
-require 'rake'
-include Rake::DSL
-SpritesRailsTestApp::Application.load_tasks
