+++ spec/fixtures/rails_project1/config/application.rb
@@ -1,12 +1,9 @@
-require File.expand_path('../boot', __FILE__)
-
 # Pick the frameworks you want:
 require "active_record/railtie"
 require "action_controller/railtie"
 # require "action_mailer/railtie"
 # require "active_resource/railtie"
 require "sprockets/railtie"
-require 'sprites/rails/autoload'
 
 if defined?(Bundler)
   # If you precompile assets before deploying to production, use this line
