--- spec/fixtures/rails_project1/config/environment.rb
@@ -1,7 +1,0 @@
-# Load the rails application
-require File.expand_path('../application', __FILE__)
-
-require 'rake'
-include Rake::DSL
-
-SpritesRailsTestApp::Application.initialize!
